Move bot-related files to its own package
